move all files in separate folder
