--- Didukh_Lopatetskiy_FI-13_14/README.md
@@ -1,64 +1,0 @@
-# InvertedIndexLib
-for now u should use only first 5 steps))))))))))))))
-i ll change it soon
-
-## Build and Usage
-
-Follow these steps to build and run LongArithmeticLib, including tests and examples:
-
-1. Clone the LongArithmeticLib repository to your local machine:
-```bash
-git clone <link>
-```
-
-2. Navigate to the project directory:
-```bash 
-cd 
-```
-
-3. Create a build directory:
-```bash
-mkdir build
-cd build
-```
-   
-4. Generate the build system using CMake:
-```bash
-cmake ..
-```
-
-5. Build the library:
-```bash
-make
-```
-   
-Run Tests
-
-1. After building the library, you can run the tests using the following command:
-```bash
-will add 
-```
-   
-Run Example Program
-
-1. Once the library is built, you can run the example program by executing:
-```bash
-./example/main
-```
-   This program demonstrates the usage of LongArithmeticLib for long arithmetic operations.
-
-Usage
-
-To use LongArithmeticLib in your own project, follow these steps:
-
-1. Include the necessary header files in your C++ code:
-```c++
-#include "invertedIndex.hpp"
-```
-
-2. Link your project with the LongArithmeticLib library:
-```cmake
-target_link_libraries(InvertedIndex InvertedIndex)
-```
-
-3. Use the provided functions for long arithmetic operations in your code.
