+++ Didukh_Lopatetskiy_FI-13_14/include/invertedIndex.hpp
@@ -9,11 +9,12 @@
 #include <regex>
 
 namespace {
-    std::regex insertPattern("^\\s*INSERT\\s+([a-zA-Z][a-zA-Z0-9_]*)\\s+\\{(.+?)\\};", std::regex::icase);
+    std::regex insertPattern(R"(^\s*INSERT\s+([a-zA-Z][a-zA-Z0-9_]*)\s+\{((?:\s*[-+]?[0-9]+\s*,\s*)*[-+]?[0-9]+)\s*};)", std::regex::icase);
+
     std::regex createPattern("^\\s*CREATE\\s+([a-zA-Z][a-zA-Z0-9_]*)\\s*;\\s*$", std::regex::icase);
     std::regex printIndexPattern("^\\s*PRINT_INDEX\\s+([a-zA-Z][a-zA-Z0-9_]*)\\s*;\\s*$", std::regex::icase);
-    std::regex containsPattern("^\\s*CONTAINS\\s+([a-zA-Z][a-zA-Z0-9_]*)\\s+\\{(.+?)\\};", std::regex::icase);
-    std::regex searchPattern("^\\s*SEARCH\\s+([a-zA-Z][a-zA-Z0-9_]*)\\s+\\{(.+?)\\};", std::regex::icase);
+    std::regex containsPattern(R"(^\s*CONTAINS\s+([a-zA-Z][a-zA-Z0-9_]*)\s+\{((?:\s*[-+]?[0-9]+\s*,\s*)*[-+]?[0-9]+)\s*};)", std::regex::icase);
+    std::regex searchPattern(R"(^\s*SEARCH\s+([a-zA-Z][a-zA-Z0-9_]*)\s+\{((?:\s*[-+]?[0-9]+\s*,\s*)*[-+]?[0-9]+)\s*};)", std::regex::icase);
 
     std::regex identifierPattern("[a-zA-Z][a-zA-Z0-9_]*");
     std::regex numberPattern("\\s*([^,]+)\\s*,?");
@@ -22,8 +23,18 @@ namespace {
 class Parser {
 private:
     std::vector<std::string> tokens;
-private:
 
+private:
+    // processes commands with this pattern: <command_name> <collection_name>;
+    void processCommand(const std::smatch& match, const std::string& commandName);
+    // processes commands with this pattern: <command_name> <collection_name> {val1, val2, ... };
+    void processCommandWithValues(const std::smatch& match, const std::string& commandName);
+   
+    void processCreateCommand(const std::smatch& match);
+    void processInsertCommand(const std::smatch& match);
+    void processPrintIndexCommand(const std::smatch& match);
+    void processSearchCommand(const std::smatch& match);
+    void processContainsCommand(const std::smatch& match);
 public:
     // add method to get string from user 
     void lexer(std::string inputString);
@@ -44,7 +55,8 @@ class Collections {
 private:
     std::map<std::string, Collection> collections;
     Parser parser;
-
+    
+    std::set<int> getSetFromTokens(const std::vector<std::string>& tokens);
 public:
     void createCollection(const std::string& collecntionName);
     void insertSet(const std::string& collecntionName, const std::set<int>& set);
