+++ Didukh_Lopatetskiy_FI-13_14/src/invertedIndex.cpp
@@ -1,101 +1,76 @@
 #include "invertedIndex.hpp"
 
+void Parser::processCommand(const std::smatch& match, const std::string& commandName) {
+    std::string collectionName = match[1];
+
+    if (std::regex_match(collectionName, identifierPattern)) {
+        tokens.push_back(commandName);
+        tokens.push_back(collectionName);
+    } else {
+        std::cout << "Error: Invalid collection name." << std::endl << std::endl;
+    }
+}
+
+void Parser::processCreateCommand(const std::smatch& match) {
+    processCommand(match, "CREATE");
+}
+
+void Parser::processPrintIndexCommand(const std::smatch& match) {
+    processCommand(match, "PRINT_INDEX");
+}
+
+void Parser::processCommandWithValues(const std::smatch& match, const std::string& commandName) {
+    std::string collectionName = match[1];
+    std::string values = match[2];
+
+    if (std::regex_match(collectionName, identifierPattern)) {
+        tokens.push_back(commandName);
+        tokens.push_back(collectionName);
+
+        std::sregex_iterator it(values.begin(), values.end(), numberPattern);
+        std::sregex_iterator end;
+
+        while (it != end) {
+            std::smatch numberMatch = *it;
+            std::string numberToken = numberMatch[1];
+            tokens.push_back(numberToken);
+            ++it;
+        }
+    } else {
+        std::cout << "Invalid collection name." << std::endl << std::endl;
+    }
+}
+
+void Parser::processInsertCommand(const std::smatch& match) {
+    processCommandWithValues(match, "INSERT");
+}
+
+void Parser::processSearchCommand(const std::smatch& match) {
+    processCommandWithValues(match, "SEARCH");
+}
+
+void Parser::processContainsCommand(const std::smatch& match) {
+    processCommandWithValues(match, "CONTAINS");
+}
+
 void Parser::lexer(std::string inputString) {
     clearTokens();
     std::smatch match;
 
-    if(std::regex_match(inputString, match, createPattern)) {
-        std::string commandName = "CREATE";
-        std::string collectionName = match[1];
+    if(std::regex_match(inputString, match, createPattern))
+        processCreateCommand(match);
 
-        if (std::regex_match(collectionName, identifierPattern)) {
-            tokens.push_back(commandName);
-            tokens.push_back(collectionName);
-        } else {
-            std::cout << "Invalid collection name." << std::endl << std::endl;
-        }
-    }
+    if (std::regex_match(inputString, match, insertPattern)) 
+        processInsertCommand(match);
 
-    if (std::regex_match(inputString, match, insertPattern)) {
-        std::string commandName = "INSERT";
-        std::string collectionName = match[1];
-        std::string values = match[2];
+    if (std::regex_match(inputString, match, printIndexPattern))
+        processPrintIndexCommand(match);
 
-        if (std::regex_match(collectionName, identifierPattern)) {
-            tokens.push_back(commandName);
-            tokens.push_back(collectionName);
-        
-            std::sregex_iterator it(values.begin(), values.end(), numberPattern);
-            std::sregex_iterator end;
-            
-            while (it != end) {
-                std::smatch match = *it;
-                std::string numberToken = match[1];
-                tokens.push_back(numberToken);
-                ++it;
-            }
-        } else {
-            std::cout << "Invalid collection name." << std::endl << std::endl;
-        }
-    }
+    if (std::regex_match(inputString, match, searchPattern))
+        processSearchCommand(match);
 
-    if (std::regex_match(inputString, match, printIndexPattern)) {
-        std::string commandName = "PRINT_INDEX";
-        std::string collectionName = match[1];
-
-        if (std::regex_match(collectionName, identifierPattern)) {
-            tokens.push_back(commandName);
-            tokens.push_back(collectionName);
-        } else {
-            std::cout << "Invalid collection name." << std::endl << std::endl;
-        }
-    }
-
-    if (std::regex_match(inputString, match, searchPattern)) {
-        std::string commandName = "SEARCH";
-        std::string collectionName = match[1];
-        std::string values = match[2];
-
-        if (std::regex_match(collectionName, identifierPattern)) {
-            tokens.push_back(commandName);
-            tokens.push_back(collectionName);
-        
-            std::sregex_iterator it(values.begin(), values.end(), numberPattern);
-            std::sregex_iterator end;
-            
-            while (it != end) {
-                std::smatch match = *it;
-                std::string numberToken = match[1];
-                tokens.push_back(numberToken);
-                ++it;
-            }
-        } else {
-            std::cout << "Invalid collection name." << std::endl << std::endl;
-        }
-    }
-
-    if (std::regex_match(inputString, match, containsPattern)) {
-        std::string commandName = "CONTAINS";
-        std::string collectionName = match[1];
-        std::string values = match[2];
-
-        if (std::regex_match(collectionName, identifierPattern)) {
-            tokens.push_back(commandName);
-            tokens.push_back(collectionName);
-        
-            std::sregex_iterator it(values.begin(), values.end(), numberPattern);
-            std::sregex_iterator end;
-            
-            while (it != end) {
-                std::smatch match = *it;
-                std::string numberToken = match[1];
-                tokens.push_back(numberToken);
-                ++it;
-            }
-        } else {
-            std::cout << "Invalid collection name." << std::endl << std::endl;
-        }
-    }
+    if (std::regex_match(inputString, match, containsPattern))
+        processContainsCommand(match);
 }
 
 std::vector<std::string> Parser::getTokens() {
@@ -103,87 +78,91 @@ std::vector<std::string> Parser::getTokens() {
 }
 
 void Parser::clearTokens() {
-    tokens.clear();
+    if(tokens.size())
+        tokens.clear();
 }
 
+std::set<int> Collections::getSetFromTokens(const std::vector<std::string>& tokens) {
+    std::set<int> setToInsert;
+    for(int j = 2; j < tokens.size(); j++) {
+        try{
+            setToInsert.emplace(std::stoi(tokens.at(j)));
+        } catch(const std::invalid_argument& e) {
+        
+        }
+    } 
+
+    return setToInsert;
+}
 
 void Collections::parse(const std::string& inputString) {
     parser.lexer(inputString);
     auto tokens = parser.getTokens(); 
+    
+    if(tokens.size()) {
+        std::string collectionName = tokens.at(1);
 
-    for(int i = 0; i < tokens.size(); i++) {
-        int collectionNamePosition = 0;
+        if(tokens.at(0) == "INSERT") 
+            insertSet(collectionName, getSetFromTokens(tokens));
 
-        if(tokens.at(i) == "INSERT") {
-            std::string collectionName = tokens.at(i + 1);
-            std::set<int> setToInsert;
+        if(tokens.at(0) == "CREATE")
+            createCollection(collectionName);
 
-            for(int j = collectionNamePosition + 1; j < tokens.size(); j++) {
-                try{
-                    setToInsert.emplace(std::stoi(tokens.at(j)));
-                } catch(const std::invalid_argument& e) {
-                
-                }
-            } 
-            insertSet(collectionName, setToInsert);
-            break;
-        }
+        if(tokens.at(0) == "PRINT_INDEX") 
+            printCollectionIndex(tokens.at(1));
 
-        if(tokens.at(i) == "CREATE") {
-            createCollection(tokens.at(i + 1));
-            break;
-        }
+        if(tokens.at(0) == "SEARCH") 
+            searchInCollection(collectionName, getSetFromTokens(tokens));
 
-        if(tokens.at(i) == "PRINT_INDEX") {
-            printCollectionIndex(tokens.at(i + 1));
-            break;
-        }
-
-        if(tokens.at(i) == "SEARCH") {
-            std::string collectionName = tokens.at(i + 1);
-            std::set<int> setToInsert;
-
-            for(int j = collectionNamePosition + 1; j < tokens.size(); j++) {
-                try{
-                    setToInsert.emplace(std::stoi(tokens.at(j)));
-                } catch(const std::invalid_argument& e) {
-                
-                }
-            } 
-            searchInCollection(collectionName, setToInsert);
-            break;
-        }
-
-        if(tokens.at(i) == "CONTAINS") {
-            std::string collectionName = tokens.at(i + 1);
-            std::set<int> setToInsert;
-
-            for(int j = collectionNamePosition + 1; j < tokens.size(); j++) {
-                try{
-                    setToInsert.emplace(std::stoi(tokens.at(j)));
-                } catch(const std::invalid_argument& e) {
-                
-                }
-            }
-            containsCollection(collectionName, setToInsert);
-            break;
-        }
+        if(tokens.at(0) == "CONTAINS") 
+            containsCollection(collectionName, getSetFromTokens(tokens));
+    } else {
+        std::cout << "Error: Invalid command syntaxis, try again.\n";
     }
 
 }
 
-
 void Collections::createCollection(const std::string& collectionName) {
-    collections[collectionName] = Collection();
-    std::cout << "Collection " << collectionName << " has been created." << std::endl;
+    if (collections.find(collectionName) != collections.end()) {
+        std::cout << "Error: Creating collection with existing name '" << collectionName << "'" << std::endl;
+    } else {
+        collections[collectionName] = Collection();
+        std::cout << "Collection '" << collectionName << "' has been created." << std::endl;
+    }
 }
 
 void Collections::insertSet(const std::string& collecntionName, const std::set<int>& set) {
     collections[collecntionName].insert(set);
+    std::cout << "Set has been added to " << collecntionName << std::endl;
 }
 
 void Collections::printCollectionIndex(const std::string &collectionName) {
+    std::cout << "Printing " << collectionName << "collection...\n";
     collections[collectionName].print_index();
+}
+
+void Collections::searchInCollection(const std::string& collectionName, const std::set<int>& set) {
+    if (collections.find(collectionName) == collections.end()) {
+        std::cout << "Error: Searching in collection that doesn't exist '" << collectionName << "'" << std::endl;
+    } else {
+        std::cout << "search in collection will be executed with this input: " << collectionName << ", ";
+        for(const auto& num : set)
+            std::cout << num << " ";
+
+        std::cout << std::endl;
+    }
+}
+
+void Collections::containsCollection(const std::string &collectionName, const std::set<int> &set) {
+    if (collections.find(collectionName) == collections.end()) {
+        std::cout << "Error: Contains method called in collection that doesn't exist '" << collectionName << "'" << std::endl;
+    } else {
+        std::cout << "search in collection will be executed with this input: " << collectionName << ", ";
+        for(const auto& num : set)
+            std::cout << num << " ";
+
+        std::cout << std::endl;
+    }
 }
 
 void Collection::insert(const std::set<int>& set) {
@@ -192,24 +171,6 @@ void Collection::insert(const std::set<int>& set) {
     for(auto it = set.begin(); it != set.end(); ++it) {
         invertedIndex[*it].push_back("set" + std::to_string(sets.size()));
     }
-}
-
-void Collections::searchInCollection(const std::string& collectionName, const std::set<int>& set) {
-    std::cout << "search in collection will be executed with this input: " << collectionName << ", ";
-    
-    for(const auto& num : set)
-        std::cout << num << " ";
-    
-    std::cout << std::endl;
-}
-
-void Collections::containsCollection(const std::string &collectionName, const std::set<int> &set) {
-    std::cout << "contains will be executed with this input: " << collectionName << ", ";
-    
-    for(const auto& num : set)
-        std::cout << num << " ";
-    
-    std::cout << std::endl;
 }
 
 void Collection::print_index() {
