--- examples/CMakeLists.txt
@@ -1,7 +1,0 @@
-add_executable(main
-    main.cpp
-)
-
-target_link_libraries(main
-    InvertedIndex
-)
